+++ package.json
@@ -34,11 +34,19 @@
     "webpack-dev-server": "^3.11.0"
   },
   "dependencies": {
+    "bcryptjs": "^2.3.0",
+    "dotenv": "^8.2.0",
+    "cookie-parser": "^1.4.1",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
+    "react-hot-loader": "^4.6.3",
     "@material-ui/core": "^4.10.1",
     "@material-ui/icons": "^4.9.1",
     "axios": "^0.19.2",
     "cors": "^2.8.5",
     "express": "^4.17.1",
+    "express-session": "^1.13.0",
+    "node-fetch": "^2.3.0",
     "pg": "^8.2.1",
     "postgres": "^1.0.2",
     "react": "^16.13.1",
