+++ src/server/routes/authRouter.js
@@ -1,0 +1,15 @@
+// const express = require('express');
+// const authController = require('../controllers/authController');
+// const router = express.Router();
+
+// router.get('/:name', authController.getVerification, authController.getUserInformation (req, res) => {
+//   return res.status(200).json({user: res.locals.userInformation});
+// });
+
+// router.post('/:name', authController.postUserInformation, authController.postVerification, 
+//   (req, res) => {
+//     return res.status(200).json({user: res.locals.userInformation});
+//   }
+// );
+
+// module.exports = router;
